Support swift and oslo milestone releases
For intermediary releases of Swift, first a RC tag is pushed on master
and bugs get fixreleased (using swiftrc.sh) then we use milestone.sh
in SKIPBUGS mode to push the final tag and upload the resulting tarball
without affecting bugs.

oslo-incubator does not produce tarballs, but we still want tags and
mark the milestone "released". We use milestone.sh in SKIPTARBALL mode
to just skip tarball wait and call upload_release.py with the --nop
argument to skip tarball upload.

oslo.messaging is not tagged at milestones, but we still want to mark
the milestone "released". We use milestone.sh in SKIPTARBALL and SKIPTAG
mode to skip tagging completely, skip tarball wait and call
upload_release.py with the --nop argument to skip tarball upload.

Change-Id: I7e670d2dd5052460a593a06e82dc212d176a395d