Check for tardiff and add site to README.
When I rad milestone.sh the first time, one error I hit was that I
didn't have tardiff installed (and obviously didn't read the prereqs
in the README).  Note where you can find tardiff in the README and
also check for tardiff to be available as a sanity check before using
it.

Change-Id: Iafcbd173d3cd06af5873686c38af5aa3bd780e16