Skip tagging if tag exists
If the repository for the project already has a tag with the given name,
ask the user whether they want to continue or not.

This change also includes a drive-by fix for the case where the
launchpad project specified on the command line does not match the
repository name (as with python-stevedore and python-cliff).

Change-Id: I668bd7d57c7406c9ace2467f468fe49fc7cb8b29