Allow spec2bp.py to work for Oslo projects
Oslo projects share a common specs repository, but since we are in a
different Launchpad project group from the rest of OpenStack the URL for
the blueprints generated by default does not match. This change looks
for "oslo" as the project name that we parse from the spec file and
converts it to "oslo-incubator" by default. To handle specs for other
Oslo libraries, a new option is provided to explicitly set the Launchpad
project name.

Change-Id: I27c76b93a55856ef4fac48debf706a4d152a740a