Fix use of actual vs. expected
In https://review.openstack.org/#/c/147256/5/sanity_check_version.py,cm
it was pointed out that the use of "actual" and "expected" for variable
names seemed backwards here. Reverse their use, and add a missing test
to verify that the old check for a minor version being incremented
too far was broken but is now fixed by the name reversal.

Change-Id: Ia9a2139ac3a6bf4c289b9f90356993beaa7961a7