Support neutron advanced services
The recently split neutron-*aas need special treatment in release
scripts. The extra deliverables need to be uploaded to the same
Launchpad milestone page, and bugs should only be closed once (all
projects share the same Launchpad bug pages).

To this effect, we distinguish in the scripts the Launchpad concept
of project from the git project name as used in tarballs and skip
parts as needed. We also introduce an optional --deliverable option
to upload_release.py so that it can use different values for both.

Change-Id: I8a4db8b09a3b9d480a767ea0ce94373b4528d1a2