Add script for creating stable releases of libraries
Libraries do not use proposed branches, and go directly to creating
stable branches using a pre-tagged release version. This new script
makes that easy to coordinate and ensures that the desired version also
exists in launchpad as a released milestone.

The new script also uses some logic that was present in
release_library.sh, so refactor that to start a library of functions for
shell scripts to share.

Change-Id: I04de68be189944eb5f611ca43f452d6bd3ba911d