Update rccut and rcdelivery to match Kilo process
For Kilo we switched to using stable/* instead of proposed/* as the
pre-release branch. Update scripts to match that.

We also stop waiting for tarballs to be built and compare them to
previous branch tarballs -- we can't rely on that test given how much
time it takes to get branch tarballs generated those days.

Change-Id: I46f069ccb5e5a6baa9c768f7d5e1c7d1237e6546