Add adjust_blueprints.py script
adjust_blueprints.py is meant to be run just before milestone tagging,
to target recently-implemented blueprints to the milestone (and set the
series goal), and clear the milestone target from any incomplete
blueprint that may still have it.

It is part of the new "report rather than predict" approach to blueprint
tracking.

Change-Id: I84d185da32fa0c0caeadd29ad9a7fc010953d080