Close milestones for post-version releases
In the process of cleaning up the alpha version handling the call to
close the milestone was removed. The old call was to upload_milestone.py
with --nop to cause the upload to be skipped. Rather than restoring it
directly, create a new script that only closes the milestone and use
that.

Change-Id: Idf7bef35e290d2888800d33aba17ca744c90bfc9