launchpad: Add script to clean up bug inconsistencies
The bugs in launchpad get sometimes in an inconsistent state. For
example there are bugs which have an assignee, but are set to 'New'.
Right now it's part of the bug triage (see task 3 at [1]) for every
project to solve those inconsistencies. This can be done automatized.
This patch set adds a script which does the cleanup automatically.

For example, to set all 'New' bugs in 'Nova' which have an assignee
to 'In Progress', the script has to be executed like this:

    python bug_cleanup.py nova --newbugs

By doing it this way it is also more "set in stone" how a consistent
bug should look like.

[1] https://wiki.openstack.org/wiki/BugTriage

Change-Id: If78298058c3f7bbd0e5c27d0d87ccf13af675a1f