Be safer with git refs.
HEAD is repo-local and can be confused when talking about e.g.
branches or existing repositories. Rather than giving HEAD as an
example, give a reference to gerrit/master which is a much more likely
actual revision to want to tag. Also tweak pull_repo to make master ==
gerrit/master, so that if someone does use HEAD its going to be what
they might expect.

Change-Id: Id7809f4e51a41e362d8feef2919411821ecf611a