Script to release cycle-with-intermediary
Script used for Swift 2.4.0 release, should be usable for any
cycle-with-intermediary items. Will probably be refactored at
some point with other release tools, but it's a good reference
to keep outside of my hard disk. This script currently only
tags on master, so it can't be used for release branch (or stable
branch) releases.

This also removes swiftrc.sh which was previously used to tag
RCs on intermediary releases of Swift. We don't do RCs on
cycle-with-intermediary items, so no point in keeping this one
around.

Change-Id: I3722e9cf222ed432a67f980bd4cd62ab83b14670