Release tagging automation
This change implements the Mitaka-style release process.

The new release_from_yaml.py is called (manually for the moment)
when a deliverable release request is merged in openstack/releases.

It calls release.sh to tag repositories in that deliverable:
- applying the tag if it does not exist
- adding comments referencing the release to Launchpad for bugs
  mentioned ("Closes-Bug") in git commit messages since last tag

add_comment.py is the launchpadlib script handling that last action.

Change-Id: I40b5f5957ff765d2c582d126b940260bb47547e8