make release_from_yaml.py figure out the deliverable files for us
Change the args to release_from_yaml.py so that if no deliverable files
are provided it looks at the git history of the releases repository to
find the files changed in the most recent commit.

Call release.sh for each of the files given (or discovered) instead of
just a single file.

Change-Id: I7dc392df1d6b9c7e4e1d4d85c9690ee9480e8c31