use plain-text file for launchpad credentials
Authenticating to use launchpad now defaults to an encrypted keyring,
which requires the user to enter a keyring password
interactively. Switch to using a plain-text file explicitly to bypass
this to improve automation. The hard-coded path to the file corresponds
to where it was created by default before the switch to keyring, and we
can add an option to specify the filename later to make the scripts work
in the CI system.

Change-Id: Ifed22f7c178b1e62e7b5e95d0f8810633ba46fb1