start sending release announcements to openstack-dev by default
The Nov/Dec 2015 mailing list thread discussing where release
announcements should go concluded that announcements about library
releases should go to openstack-dev instead of openstack-announce [1].
Since most of our releases are for libraries, change the default for the
announcement. We will have more work to do to update the announcement
script to set an appropriate "to address" based on the deliverable type.

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-December/082171.html

Change-Id: I777f8ab7eafd661932655d5e67de3c494caa110f