refactor list_deliverable_changes.py prior to move to project-config
When we move the release tagging scripts to the project-config
repository, they will need to have minimal dependencies and shouldn't
need to import other local modules. Move the gitutils code needed by
list_deliverable_changes.py into the file, and move the file out of the
cmds subdirectory to be a regular script instead.

Change-Id: I7076d45be04366b95b3d662ffc181373339ad6c6