refactor script for commenting on launchpad bugs prior to move to project-config
When we move the release tagging scripts to the project-config
repository, they will need to have minimal dependencies and shouldn't
need to import other local modules. Move the launchpad code needed by
launchpad_add_comment.py into the file, and move the file out of the
cmds subdirectory to be a regular script instead.

Change-Id: Ie004c2a14100611d4b7c1604dcf7f465a6f0246c