Added tool to filter out bugs of specified importance
Some LP bugs are not really bugs, in a sense that they are actual
features required. Those fixes are not applicable for stable backports,
and hence should be ignored when determining candidates for stable
branches.

Also, for some older stable branches, we may want to filter out those
bugs that do not satisfy importance criteria [e.g. for stable/kilo we
require that only High+ bugs are to be backported].

This script is useful if you already have a list of bugs to consider for
backports. F.e. you can generate the list with bugs-fixed-since.py tool.

Common use case would be:

./bugs-fixed-since.py --start=<prev-hash> | ./lp-filter-bugs-by-importance.py neutron

That would give you the list of bug numbers that were fixed in neutron
master since <prev-hash> and that are not of Wishlist importance.

Change-Id: I07e8d50b826aff35df2c2360d09bf029c036b070