change email topic tags based on summit feedback
At the feedback session at the Newton summit, we had feedback that
publishing release announcements under the "[release]" topic tag
cluttered up the mail being filtered by folks who need to listen for
information from the release team. Similarly, adding [stable] as a tag
to email related to stable branches makes far more email show up for
folks who need to watch for instructions from the stable team.

This patch changes "release" to "newrel" and removes the stable tag from
the subject line entirely. The body of the message still says when the
release is part of a stable series.

Change-Id: I5f45b9650df2edf5466cc9ea8c9dbce9c667c34c
Signed-off-by: Doug Hellmann <doug@doughellmann.com>