fix the way we detect the previous version in release.sh
Use the SHA as a reference, instead of making assumptions about branch
names, when determining the previous version number. This gives us a
more accurate previous version in cases like openstack-ansible, where
the branch names for older branches do not follow the assumed
pattern. It also should fix issues with retagging a commit with a new
tag.

Change-Id: Ieff91e799767dfe9e198c0a2d23d93174ddaaa26
Signed-off-by: Doug Hellmann <doug@doughellmann.com>