announce: make metatag optional
This allows the script to work even if the metatag are not present in
the git tag. It will then send a simple email announcing the release,
without any series information, with just the release information
between the PREVIOUS_VERSION and VERSION.

This will allow independent projects to be managed by this script
without going through openstack/releases.

Change-Id: I9881ad83d25a53db2d2c314d1accf847bb79aa8d