add script to expire old bug reports
Many OpenStack projects face the issue to have more bug reports than
they can handle with the resources they got. The chances, that the
same setting which leads to the observed issue is still there, is quite
often unlikely. It's better to be explicit and expire old bug reports
which will not get any attention any more.

Change-Id: I3c9c72b528b090647211262d2f0030d66be9857c