release_notes: support multiple README formats
Add support for md in the list of formats readabl in the
release_notes.py script.
Some projects (Puppet OpenStack modules) use markdown for README.

If README.rst nor README.md is not found, raise an error before trying
to read the file to grab informations that we need later.

Change-Id: Ia3c0cd571cc1745926b41f4050f04133b53ac77c