fix release notes for final releases
We want release announcements for final releases, after release
candidates, to include all of the changes from the release. In a perfect
world, the tag history of the releases would be such that we could ask
git for the previous release. Unfortunately, the tag for the final
release from the previous series is not on the same branch where we are
tagging the final of the new release. It used to be merged in, but at
the wrong point. Now it isn't even being merged in.

Rather than trying to deduce the right start point in the announce
script, this change makes it an optional piece of the tag metadata. When
present, the diff-start value is used. When not present, the old
behavior of using the previous tag is retained.

Given our practice of creating new stable branches on the first release
candidate tag, we can also update propose-final-releases to
automatically set the diff-start value for new final releases to point
to the first release candidate of the last release series. If that
automatic value is wrong, it can be adjusted manually before the final
tags are applied.

See related work in I4dff3a57f4d0c8f625e789ec19683c47983b7877

Change-Id: I79c3ff7cc07300af5336a0e9f232f8802d4b4c4a
Signed-off-by: Doug Hellmann <doug@doughellmann.com>