Do not abort if README file not found
Currently if README.{md,rst} is not found in the source repo announce.sh
raises a RuntimeError and aborts.  This means that the email template
isn't written.  This ultimately overall job failing as the email can't
be sent.

A recent example of this is at:
    http://logs.openstack.org/03/03c4534b0e1ba2a6c4ad0aaa471070fd53b2ae73/release/xstatic-bootstrap-scss-announce-release/288f27d/console.html#_2016-08-18_18_21_11_191756

The README file us used to get the bug and source urls.  The jinja
templates are setup to function correctly if these are empty strings.
So demote this situation to a warning and announce the release.

Change-Id: Ie492191a79b903a8079fa64ef11a71fbb4fb4903