update release dashboard output for newton
We no longer distinguish between managed and unmanaged projects using a
separate tag, so remove that level from the data structure and output.

Add a few empty columns based on what we ended up adding by hand for
Mitaka.

Add a link to the release notes, when available, so we can track which
projects have provided the link to their series-specific page.

Change-Id: I6eff1bb2d70a463d86ae5b0b3fdc1f2d9c904ea9
Signed-off-by: Doug Hellmann <doug@doughellmann.com>