Fix update_upper_constraints removing additional ' ' char
In update_upper_constraints the sed substitution matches on a ' ' char after
the UPPER_CONSTRAINTS_FILE consumes.  This creates bad changes such as [1].
Compare the manual correction at [2].

I know that make_stable_branches is being moved to project-config but
the shell function remains in release-tools so shoudl be fixed.

[1] https://review.openstack.org/#/c/364973/1/tox.ini@17
[2] https://review.openstack.org/#/c/364973/1..2/tox.ini,unified

Change-Id: I373612254fa7b4afdce0b76a98ad3a88d1f77911