make the script for proposing final releases more flexible
We have a few deliverables in newton for which there are no release
candidates, but there are other pre-releases (betas) from which we want
to cut the final release. Make the logic for finding them more flexible.

The result will still assume that we want to diff from the rc1, which
won't exist, but those values were always just suggestions so we will
need to fix them by hand anyway.

Change-Id: I7fddbc5fec28eb1fe27ccfffa6e9a5126bc51dfd
Signed-off-by: Doug Hellmann <doug@doughellmann.com>