do not use locale version of admonition labels
The admonition labels from the Sphinx code base are translated, but we
don't initialize the translation machinery when we convert the release
notes to text because we only publish the original text version. The
labels default to capitalized versions, so just use title() to produce
something similar without going through the un-initialized locale code.

This fixes the traceback we saw when publishing the newton release
announcement for nova:
http://logs.openstack.org/ad/ad68ee297842d142749877ccc66a6b677b576279/release/nova-announce-release/b6ca7ae/console.html

Change-Id: I29906fd0af2a507e3131923a659b7448d293dbc7
Signed-off-by: Doug Hellmann <doug@doughellmann.com>