ignore unrecognized inline directives
When we see a directive we do not understand, skip it instead of
reporting an error. This avoids embedding system error messages in the
output at the expense of sometimes leaking rst directives.

See the murano-dashboard 3.0.0 release announcement for an example of
the error this
addresses. http://lists.openstack.org/pipermail/openstack-announce/2016-October/001748.html

Change-Id: I28dd334b7d326e13e7aaf7cd9c0a160ae1e9fa09
Signed-off-by: Doug Hellmann <doug@doughellmann.com>