fix tools for listing unreleased changes
The tool for listing repositories is now in the openstack/releases repo,
so update list_library_unreleased_changes.sh to take a path to a copy of
that repo.

The release-notes command needs a name for the thing being queried, so
pass it the base repository name. That may not be correct in all cases,
but we don't care that it's correct because we're not publishing the
information as authoritative.

Depends-On: Iaf0cd77237c0b1892aaa0864eb356012c57f036c
Change-Id: Ib9dcbbc1c3325ce523a946cca750333add51eedb
Signed-off-by: Doug Hellmann <doug@doughellmann.com>