Add support for mail servers that require a login
When running announce.sh from my local developer workstation I cannot
connect to lisst.openstack.org.  I know my ISP filters outbound traffic
to port TCP:25 but there may be additional reasons.

This chnage enhances the --server option to allow you to provide
username, password hostname and TCP port with which to contact a mail
server.

This is not a complete solution but it makes it much easier for me to
work locally if needed.

Change-Id: I1a0f03341d1962a30c0737b59d32e8e68c28644b