fix mail sending
Correct the logic related to parsing the server address so that the
default is not overwritten if there is no @ in the address.

Fix handling of a missing password by allowing an empty string to mean
no password was given.

Move the TO parsing out of the try/finally block so that errors in
parsing the message are not swallowed by errors when we quit a server
connection that wasn't opened properly.

Change-Id: I73a24a4d87c0f4a5aabd302d8a7d2884e5513c31
Signed-off-by: Doug Hellmann <doug@doughellmann.com>