Check CI environment for project to announce
In the event that we're releasing a non-Python project or the
project otherwise lacks a usable setup.py, first attempt to infer
the project name from the $ZUUL_PROJECT environment variable if set
before falling back on assuming the project name is the same as the
name of the current working directory. This is particularly
important when running under CI systems like Zuul, which will name
the current working directory after the job rather than after the
project itself.

Change-Id: I19b51846b581c5d033b08466ca09f23b23bd5e49