Set up gerrit remote manually in eol_branch.sh
When a project is retired there is not .gitreview in their repository.
If we still need to EOL old branches we need to figure out the gerrit
remote. We could get this by checking out an older branch, but to keep
it simple we just assume/guess it's review.openstack.org/$project.git

Change-Id: I779783cc3f06af28f17ee36a1f0a84df29c4177d