Fix abandoning reviews in eol_branch
The loop that was meant to abandon all open reviews for an EOL branch was
only getting the 1st one due to the way the ssh output was being read in.

Make this neater by creating an array of revisions from the query to
abandon before looping over them.

Additionally tighten the regex matching the change revision. The gerrit
query returns a lot of information, including the commit message, meaning
that any references to 'revision' in a commit message could be erroneously
matched.

Change-Id: I1dd3a94bec45aaeae9c37af1fd7b18905f0401ef