fix the repo name detection when we don't have an origin set
Look at the directory we're given as the location of the code when we
can't find the repo name using the git remote.

This should only be needed until we move more metadata like the team
owner and some of these name values into the git commit message and
pass them from the tagging job to the announce job using that metadata
channel.

http://logs.openstack.org/08/08e872a2fad9734446030aa94d064049a0db371c/release/announce-release/864dea0/job-output.txt.gz#_2017-10-19_15_31_21_974136

Change-Id: I139d8177e68d2c3c2fd1e06d7dfb5baaedc220c3
Signed-off-by: Doug Hellmann <doug@doughellmann.com>